Synchronize DMS Website with Repository
Contains latest About-Us Changes, and Join-the-Team Changes on the website.
Also, moved all files under a dms folder.  This makes it easier when deploying the site to AWS.  Just zip the dms folder, re-name the zipped folder to dms-teamdms-<description of change>.  Via AWS Elastic Beanstalk Environment, Upload this .zip and then choose option to Deploy.